changed cnadidatesCount to uint256 to avoid type conversion

+++ contracts/Election.sol
@@ -8,7 +8,7 @@ contract Election {
         uint voteCount;
     }
     Candidate[] public candidates;
-    uint16 public candidatesCount;
+    uint256 public candidatesCount;
     
     mapping(address => bool) public voters;
